+++ src/ritas/cli.py
@@ -6,14 +6,22 @@ grid file.
 
 """
 
+import logging
+from pathlib import Path
+
 import click
 
+from ritas import LOG
 from ritas.workflows import simple_workflow
 
 
 @click.command(help=__doc__)
 @click.option(
-    "--input", "-i", "infile", type=click.Path(exists=True), required=True
+    "--input",
+    "-i",
+    "infile",
+    type=click.Path(exists=True, path_type=Path),
+    required=True,
 )
 @click.option(
     "--swath-width",
@@ -24,11 +32,18 @@ from ritas.workflows import simple_workflow
     help="Width (m) of the swath, over-riding what is in the input file.",
 )
 @click.option(
-    "--output", "-o", "outfile", type=click.Path(exists=False), required=True
+    "--output",
+    "-o",
+    "outfile",
+    type=click.Path(exists=False, path_type=Path),
+    required=True,
 )
 def main(**kwargs: dict) -> None:
     """Run the command line interface for ritas."""
+    # Set up logging
+    logging.basicConfig(level=logging.INFO)
+
     infile = kwargs.pop("infile")
     outfile = kwargs.pop("outfile")
-    click.echo(f"I am about to process {infile} -> {outfile}")
+    LOG.info("I am about to process %s -> %s", infile, outfile)
     simple_workflow(infile, outfile, **kwargs)